chore: remove unnecessary forward refs and clean up imports
Imports are now alphabetized, a personal preference of mine, but can be
ignored going forward if desired. All of the uses of `@Inject(forwardRef())`
are now removed, and the multiple calls to `ConfigModule.forRoot()` are
also removed from multiple modules. This is to help get ready for the new
config implementation where each feature module asserts its own necessary
config, without having to re-call the `forRoot()` method.