feat: new config module implementation
This new imlpementation should allow for a more direct `forFeature` kind
of approach when it comes to creating validations for the configurations for
environment variables needed by each module. This should lead to a better
dev experience than the direct `@nestjs/config` implementation without
some of the headache of having to remember what providers do and do not
exist in the global, or similar, scope.

ref: #89
ref: #90